Remove test_parser_exception_on_no_url
Due to the way argparse works, after the ArgumentError exception is
raised, sys.exit is called, which assertRaises *does not like*. An
attempted fix was to use assertRaises as a context manager, but this
didn't resolve the issue.

More detail on the issue at http://stackoverflow.com/a/15692278

Because this issue does not affect the functionality of the program (I have
verified that argparse will fail when no url is given) I have elected to
simply comment out this test from the others.